pyproject.toml
[project]
name = "research-portal"
version = "0.1.0"
description = "Add your description here"
requires-python = "~=3.12.0"
dependencies = [
    "pinecone==5.4.2",
    "python-dotenv==1.0.1",
    "google-genai==1.19.0",
    "flask==3.1.1"
]
